requirements.txt
# Устанавливаем Django как основной backend-фреймворк проекта.
Django==5.2.12

# Устанавливаем библиотеку для загрузки переменных окружения из файла .env.
python-dotenv==1.1.1

# Устанавливаем aiogram для Telegram-бота.
aiogram==3.26.0

# Устанавливаем pandas для аналитики и обработки финансовых данных.
pandas==2.3.2

# Устанавливаем matplotlib для построения графиков и диаграмм.
matplotlib==3.10.6

# Устанавливаем Pillow для работы с изображениями и графическими сценариями Django.
Pillow==11.3.0